% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_sil.R
\name{avg_sil}
\alias{avg_sil}
\title{Average Silhouette}
\usage{
avg_sil(k, matrix)
}
\arguments{
\item{k}{number of cluser}

\item{matrix}{matrix to compute average sihouette}
}
\value{
Average silhouette of the observations are calculated for each k.
}
\description{
Compute the average silhouette value for the observations for each k (Description)
}
\details{
We want to determine the optimal K clusters to use. The silhouette method is used as default(Details)
}
\examples{
avg_sil(k=4,matrix= optimMat)
}
\author{
Ellis Wong
}
