% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotshort.R
\name{ggplotshort}
\alias{ggplotshort}
\title{Plot Image}
\usage{
ggplotshort(colors, matrix, inputtitle, themevals)
}
\arguments{
\item{colors}{RGB}

\item{matrix}{the image}

\item{inputtitle}{title of plots}

\item{themevals}{the theme vals for ggplot}
}
\value{
Side by side plot of original image and k clustered image
}
\description{
Plots the images (Description)
}
\details{
We use ggplot in order to visualize the original image and k clustered image (Details)
}
\examples{
ggplotshort(kColors, imgRGB, paste('K-Means Clustering of', numclust, 'Colors'), themevals)
}
\author{
Ellis Wong
}
