% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageClassifier.R
\name{imageClassifier}
\alias{imageClassifier}
\title{This function classifies the image using either a user-specified number of clusters or the silhouette method through the avg_sil function and ggplotshort function in this package in order to plot the original image and the k clustered image (Description)}
\usage{

  imageClassifier(url = "https://res.cloudinary.com/twenty20/private_images/t_low-crop/photosp/f985a04a-3f6a-4407-b3d8-722aac5eb6d8/f985a04a-3f6a-4407-b3d8-722aac5eb6d8.jpg",
  numclust = 1)
}
\arguments{
\item{url}{jpg image to classify}

\item{numclust}{number of clusters, with 1 as default}
}
\value{
Side by side plot of original image and k clustered image
}
\description{
The default is a Pug in a floatie using the Silhouette method. This will return the original image in addition to the k-means clustered image(Details)
##' @title Image Classifier
}
\examples{
imageClassifier(url = 'https://res.cloudinary.com/twenty20/private_images/t_low-crop/photosp/f985a04a-3f6a-4407-b3d8-722aac5eb6d8/f985a04a-3f6a-4407-b3d8-722aac5eb6d8.jpg', numclust = 4)
}
\author{
Ellis Wong
}
